requirements.txt
langchain==1.2.13
langchain-core==1.2.23
langchain-groq==1.1.2
langgraph==1.1.3
langgraph-checkpoint-sqlite==3.0.3
pydantic==2.12.5
sqlalchemy==2.0.48
python-dotenv==1.2.2
streamlit==1.55.0
